requirements.txt
fastapi
uvicorn
python-docx
requests
pydantic
python-multipart
